+++ CHANGELOG.md
@@ -10,10 +10,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 - Enhanced `SeleniumUrlLoader` with running with `chromium-driver`. 
 - Running spinner and elapsed time counter. 
+- Support extra prompts.
+- Use `--data-loader` to set web driver for `SeleniumURLLoader`. Support ["firefox", "chrome", "chromium"]. 
+- Support to select the system role. 
+- Support to select output languages. 
+- [ ] Support Ollama. 
 
 ### Changed
 
 - Login button with label `Login`.
+- Switch preview/raw of response with a toggle. 
 
 
 ## [0.1.0] - 2024-06-07